motac -> tmac
