+++ apps/tmac_3s_portfolio.cc
@@ -4,7 +4,7 @@
 #include "operators.h"
 #include "parameters.h"
 #include "splitting_schemes.h"
-#include "motac.h"
+#include "tmac.h"
 #include "util.h"
 #include "MarketIO.h"
 #include <thread>
@@ -31,7 +31,7 @@ int main(int argc, char *argv[]) {
   
   // set para
   double operator_step_size = 0.0018;
-  params.motac_step_size = 0.5;
+  params.tmac_step_size = 0.5;
   params.step_size = operator_step_size;
   
   int problem_size = Q.cols();
@@ -45,7 +45,7 @@ int main(int argc, char *argv[]) {
   
 
   double start_time = get_wall_time();  
-  MOTAC(three_s, params);  
+  TMAC(three_s, params);  
   double end_time = get_wall_time();
   print_parameters(params);  
   cout << "Computing time is: " << end_time - start_time << endl;
