+++ apps/tmac_gd_least_square.cc
@@ -12,7 +12,7 @@
 #include "operators.h"
 #include "parameters.h"
 #include "splitting_schemes.h"
-#include "motac.h"
+#include "tmac.h"
 #include "util.h"
 #include <thread>
 using namespace std;
@@ -38,7 +38,7 @@ int main(int argc, char *argv[]) {
   loadMarket(b, label_file_name);
   int problem_size = A.rows();
   params.problem_size = problem_size;
-  params.motac_step_size = 0.1;
+  params.tmac_step_size = 0.1;
   int sample_size = A.cols();
   Vector x(problem_size, 0.);   // unknown variables, initialized to zero
   Vector Atx(sample_size, 0.);  // maintained variables, initialized to zero
@@ -53,9 +53,9 @@ int main(int argc, char *argv[]) {
   // Step 4. Define your operator splitting scheme
   GradientDescentAlgorithm<forward_grad_for_square_loss<Matrix> > gd(&x, forward);
 
-  // Step 6. Call the MOTAC function
+  // Step 6. Call the TMAC function
   double start_time = get_wall_time();  
-  MOTAC(gd, params);
+  TMAC(gd, params);
   double end_time = get_wall_time();  
 
   print_parameters(params);
