+++ apps/tmac_jacobi_linear_eqn.cc
@@ -12,7 +12,7 @@
 #include "operators.h"
 #include "parameters.h"
 #include "splitting_schemes.h"
-#include "motac.h"
+#include "tmac.h"
 #include "util.h"
 #include "MarketIO.h"
 #include <thread>
@@ -40,7 +40,7 @@ int main(int argc, char *argv[]) {
   loadMarket(b, label_file_name);
   int problem_size = A.rows();
   params.problem_size = problem_size;
-  params.motac_step_size = 1.;
+  params.tmac_step_size = 1.;
   Vector x(problem_size, 0.);   // unknown variables, initialized to zero
 
   
@@ -51,9 +51,9 @@ int main(int argc, char *argv[]) {
   // Step 4. Define your operator splitting scheme
   GradientDescentAlgorithm<linear_eqn_jacobi_operator<Matrix> > Jacobi(&x, J);
 
-  // Step 6. Call the MOTAC function
+  // Step 6. Call the TMAC function
   double start_time = get_wall_time();
-  MOTAC(Jacobi, params);
+  TMAC(Jacobi, params);
   double end_time = get_wall_time();    
   
    // Step 7. Print results
