+++ apps/tmac_prs_demo.cc
@@ -14,7 +14,7 @@
 #include "operators.h"
 #include "parameters.h"
 #include "splitting_schemes.h"
-#include "motac.h"
+#include "tmac.h"
 #include "util.h"
 #include "MarketIO.h"
 #include <thread>
@@ -32,7 +32,7 @@ int main(int argc, char *argv[]) {
 
   // Step 2. Load the data or generate synthetic data, define matained variables
   int problem_size = params.problem_size;
-  params.motac_step_size = 0.9;
+  params.tmac_step_size = 0.9;
 
   // define auxilary variables
   Vector x(problem_size, 10.);   // unknown variables, initialized to zero
@@ -54,9 +54,9 @@ int main(int argc, char *argv[]) {
   // Step 4. Define your operator splitting scheme
   PeacemanRachfordSplitting<First_Backward, Second_Backward> prs(&x, first, second);  
 
-  // Step 5. Call the MOTAC function  
+  // Step 5. Call the TMAC function  
   double start_time = get_wall_time();
-  MOTAC(prs, params);
+  TMAC(prs, params);
   double end_time = get_wall_time();  
 
   first(&x, &y); // do one more backward step to recover the solution
