+++ doc/source/equations.rst
@@ -1,6 +1,6 @@
 Equation
 ***********
-MOTAC currently supports solving dense and sparse linear equations with the async-parallel Jacobi method. We plan to provide support for solving nonlinear equations with the parallel asynchronous Newton-Jacobi Method, and asynchronous method for solving ODE and PDE.
+TMAC currently supports solving dense and sparse linear equations with the async-parallel Jacobi method. We plan to provide support for solving nonlinear equations with the parallel asynchronous Newton-Jacobi Method, and asynchronous method for solving ODE and PDE.
 
 .. toctree::
    :maxdepth: 1
