+++ include/algebra_namespace_switcher.h
@@ -1,5 +1,5 @@
-#ifndef MOTAC_INCLUDE_ALGEBRA_NAMESPACE_SWITCHER_H
-#define MOTAC_INCLUDE_ALGEBRA_NAMESPACE_SWITCHER_H
+#ifndef TMAC_INCLUDE_ALGEBRA_NAMESPACE_SWITCHER_H
+#define TMAC_INCLUDE_ALGEBRA_NAMESPACE_SWITCHER_H
 
 using namespace BLASAlgebra;
