+++ include/nist_spblas.h
@@ -1,5 +1,5 @@
-#ifndef MOTAC_INCLUDE_NIST_SPBLAS_H
-#define MOTAC_INCLUDE_NIST_SPBLAS_H
+#ifndef TMAC_INCLUDE_NIST_SPBLAS_H
+#define TMAC_INCLUDE_NIST_SPBLAS_H
 
 
 #endif
