+++ run_me.sh
@@ -10,7 +10,7 @@ label=""
 # function to display welcome message
 function welcome_msg {
     echo "------------------------------------------"
-    echo "|        Welcome to MOTAC demos          |"
+    echo "|        Welcome to TMAC demos          |"
     echo "------------------------------------------"
     echo "| Enter 0 to exit                        |"    
     echo "| Enter 1 to run least squares           |"
@@ -50,37 +50,37 @@ while read input </dev/tty; do
 	0 )
 	    exit 0 ;;
 	1 )
-	    app="motac_gd_ls"
+	    app="tmac_gd_ls"
 	    process_app_with_data
 	    exec="$exec_path$app -data ./data/$data -label ./data/$label -epoch $epochs -nthread $num_threads"
-	    echo "start MOTAC to solve the least squares problems."
+	    echo "start TMAC to solve the least squares problems."
 	    $exec
 	    exit 0 ;;
 	2 )
-	    app="motac_fbs_lasso"
+	    app="tmac_fbs_lasso"
 	    process_app_with_data
 	    exec="$exec_path$app -data ./data/$data -label ./data/$label -epoch $epochs -nthread $num_threads"
-	    echo "starting MOTAC to solve LASSO"
+	    echo "starting TMAC to solve LASSO"
 	    $exec
 	    exit 0 ;;
 	3 )
-	    app="motac_fbs_l2_svm"
+	    app="tmac_fbs_l2_svm"
 	    process_app_with_data
 	    exec="$exec_path$app -data ./data/$data -label ./data/$label -epoch $epochs -nthread $num_threads"
-	    echo "starting MOTAC to solve L2_SVM"
+	    echo "starting TMAC to solve L2_SVM"
 	    $exec
 	    exit 0 ;;
 	4 )
-	    app="motac_fbs_l1_log"
+	    app="tmac_fbs_l1_log"
 	    process_app_with_data
 	    exec="$exec_path$app -data ./data/$data -epoch $epochs -nthread $num_threads"
-	    echo "starting MOTAC to solve sparse logistic regression"
+	    echo "starting TMAC to solve sparse logistic regression"
 	    $exec
 	    exit 0 ;;
 	5 )
-	    app="motac_prs_demo"
+	    app="tmac_prs_demo"
 	    exec="$exec_path$app"
-	    echo "starting MOTAC to demo PRS"
+	    echo "starting TMAC to demo PRS"
 	    $exec
 	    exit 0 ;;
